+++ jsonmodel/__init__.py
@@ -1,7 +1,7 @@
 ''' A Collection of Methods for Validating JSON Structured Data '''
 __author__ = 'rcj1492'
 __created__ = '2015.12'
-__version__ = '2.8'
+__version__ = '2.9'
 __license__ = 'MIT'
 __module__ = 'jsonmodel'
 __email__ = 'support@collectiveacuity.com'